requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.0
scikit-learn==1.6.1
edazer==0.1.3
sentence-transformers==5.0.0
joblib==1.4.2
fastapi==0.115.12
pydantic==2.10.6
groq==0.30.0
python-dotenv==1.0.1



